src/modules/profileExport/kfthm-profile-export.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { KfThmProfileExportService } from './kfthm-profile-export.service';
import { KfThmProfileExportRepository } from './kfthm-profile-export.repository';
import { WorkSheet, WorkBook } from 'xlsx/types';
import * as XLSX from '@sheet/core';
import { s3 } from './../../common/common.utils';
import { KfEmailService } from '../../common/email/email.service';

jest.mock('./../../common/common.utils');
jest.mock('../../common/email/email.service');

describe('KfThmProfileExportService', () => {
    let profileExportService: KfThmProfileExportService;
    let profileExportRepository: KfThmProfileExportRepository;
    let emailService: KfEmailService;
    let module: TestingModule;

    beforeEach(async () => {
        module = await Test.createTestingModule({
            providers: [KfThmProfileExportService, KfThmProfileExportRepository, KfEmailService],
        }).compile();

        profileExportService = module.get<KfThmProfileExportService>(KfThmProfileExportService);
        profileExportRepository = module.get<KfThmProfileExportRepository>(KfThmProfileExportRepository);
        emailService = module.get<KfEmailService>(KfEmailService);
    });

    afterEach(() => {
        jest.resetAllMocks();
    });

    test('stub', () => {});
/*
    describe('profileExport', () => {
        test('should return url', async () => {
            jest.spyOn(profileExportRepository, 'createClientExport').mockResolvedValueOnce({} as any);
            // The method calls do not exist now in the service file so commenting it
            // jest.spyOn(profileExportRepository, 'getProfileExport').mockResolvedValueOnce({ row1: [{ column1: '' }] } as any);
            // jest.spyOn(profileExportService, 'convertToExcel').mockReturnValueOnce('excel');
            (s3.headBucket as any).mockReturnValueOnce({
                promise: () => {
                    return Promise.reject();
                },
            });
            (s3.createBucket as any).mockReturnValueOnce({
                promise: () => {
                    return Promise.resolve();
                },
            });
            (s3.upload as any).mockReturnValueOnce({
                promise: () => {
                    return Promise.resolve({ Key: 'mock-s3-key' });
                },
            });
            (s3.getSignedUrlPromise as any).mockResolvedValueOnce('mock-signed-url');
            (emailService.KfSendEmail as any).mockReturnValueOnce();
            const query: any = {
                userId: 123,
            };
            const body: any = {
                clientName: 'test-client-name',
                lcid: '',
            };

            const token = 'test-token';

            const result = { url: 'mock-signed-url' };

            jest.spyOn(profileExportService, 'profileExport').mockImplementation((query, body, token) => Promise.resolve(result));
            jest.spyOn(profileExportService, 'profileExport').mockResolvedValueOnce(result);
            // const res = await profileExportService.profileExport(query, body, token);
            // expect(res).toEqual(result);
        });
    });

    // The method calls do not exist now in the service file so commenting it
    // describe('columnAutoWidth', () => {
    //     // test('should return null if data.length is 0', () => {
    //     //     const res = profileExportService.columnAutoWidth({}[0], {});

    //     //     expect(res).toBeNull();
    //     // });

    //     test('should return nothing if data.length isn`t 0 and key equals KfThmProfileEnum.DESCRIPTION', () => {
    //         const res = profileExportService.columnAutoWidth([{ Description: 1 }], {});

    //         expect(res).toBeUndefined();
    //     });

    //     test('should return nothing if data.length isn`t 0 and key doesn`t equal any of KfThmProfileEnum variables', () => {
    //         const res = profileExportService.columnAutoWidth([{ Property: 1 }], {});

    //         expect(res).toBeUndefined();
    //     });
    // });

    describe('convertToExcel', () => {
        test('should convert to Excel when ws is null', () => {
            jest.spyOn(XLSX.utils, 'book_new').mockReturnValueOnce({} as any);
            jest.spyOn(XLSX.utils, 'json_to_sheet').mockReturnValueOnce({});
            jest.spyOn(XLSX.utils, 'aoa_to_sheet').mockReturnValueOnce({});
            jest.spyOn(XLSX.utils, 'book_append_sheet').mockImplementationOnce(() => '');
            // The method calls do not exist now in the service file so commenting it
            // jest.spyOn(profileExportService, 'columnAutoWidth').mockImplementationOnce(() => {});
            // jest.spyOn(profileExportService, 'styleHeaders').mockImplementationOnce(() => {});
            jest.spyOn(profileExportService, 'styleCells').mockImplementationOnce(() => {});
            jest.spyOn(XLSX, 'write').mockReturnValueOnce({});

            // The method calls do not exist now in the service file so commenting it
            // const res = profileExportService.convertToExcel({ key: { columns: {} } });
            // expect(res).toEqual(result);
        });

        test('should convert to Excel when ws isn`t null', () => {
            const dbResponse = { row1: [{ column1: '' }] };

            const workSheet: WorkSheet = {
                A2: { t: 's', v: '' },
                A1: { t: 's', v: 'column1' },
                '!ref': 'A1:A2',
            };

            const wb = {
                SheetNames: ['row1'],
                Sheets: { row1: { A2: [Object], A1: [Object], '!ref': 'A1:A2' } },
            };

            const wbExtended = {
                wb: wb,
                cellStyles: true,
                bookSST: true,
                bookType: 'xlsx',
                type: 'buffer',
            };

            jest.spyOn(XLSX.utils, 'book_new').mockReturnValueOnce({ SheetNames: [], Sheets: {} });
            jest.spyOn(XLSX.utils, 'json_to_sheet').mockReturnValueOnce(workSheet);
            jest.spyOn(XLSX.utils, 'book_append_sheet').mockImplementationOnce(() => '');

            // The method calls do not exist now in the service file so commenting it
            // jest.spyOn(profileExportService, 'columnAutoWidth').mockImplementationOnce(() => {});
            // jest.spyOn(profileExportService, 'styleHeaders').mockImplementationOnce(() => {});
            jest.spyOn(XLSX, 'write').mockReturnValueOnce(wbExtended);

            // The method calls do not exist now in the service file so commenting it
            // const res = profileExportService.convertToExcel(dbResponse);

            // expect(res).toEqual(wbExtended);
        });
    });

    describe('styleHeaders', () => {
        test('should style headers', () => {
            jest.spyOn(XLSX.utils, 'decode_range').mockReturnValueOnce({ s: { c: 0, r: 0 }, e: { c: 0, r: 1 } });
            jest.spyOn(XLSX.utils, 'sheet_set_range_style').mockImplementationOnce(() => {});
            jest.spyOn(XLSX.utils, 'encode_cell').mockReturnValueOnce('A1');
            jest.spyOn(XLSX.utils, 'sheet_set_range_style').mockImplementationOnce(() => {});

            // The method does not exist in the service
            // const res = profileExportService.styleHeaders({
            //     A2: { t: 's', v: '' },
            //     A1: { t: 's', v: 'column1' },
            //     '!ref': 'A1:A2',
            // });

            // expect(res).toBeUndefined();
        });
    });
 */
});
